Move aggregate op implementation to seperate file
